re-enable saved variable saving, not clever
By default, penlight dumps are "clever". One thing this means is
special handling of the array portion of tables, which involves a custom
ipairs implementation that blindly supplies integer keys from 1.

This is great, except that it conflicts with __index metamethods that
automatically generate values. This happened in BtWLoadouts, and caused
nontermination.

Fortunately, just disabling penlight cleverness is sufficient.